refactor: improve class name and grabbing root cause of exception
