+++ src/main/java/br/edu/ufpel/rokamoka/controller/UserRestController.java
@@ -1,12 +1,12 @@
 package br.edu.ufpel.rokamoka.controller;
 
-import br.edu.ufpel.rokamoka.context.ApiResponse;
+import br.edu.ufpel.rokamoka.context.ApiResponseWrapper;
 import br.edu.ufpel.rokamoka.dto.input.UserBasicDTO;
 import br.edu.ufpel.rokamoka.dto.output.UserResponseDTO;
 import br.edu.ufpel.rokamoka.exceptions.RokaMokaContentDuplicated;
 import br.edu.ufpel.rokamoka.security.AuthenticationService;
 import br.edu.ufpel.rokamoka.service.UserService;
-import br.edu.ufpel.rokamoka.wrapper.ApiResponseWrapper;
+import br.edu.ufpel.rokamoka.wrapper.RokaMokaController;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
@@ -18,19 +18,19 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequestMapping("/user")
 @RequiredArgsConstructor
-public class UserRestController extends ApiResponseWrapper {
+public class UserRestController extends RokaMokaController {
 
     private final UserService userService;
     private final AuthenticationService authenticationService;
 
     @PostMapping("/register")
-    public ResponseEntity<ApiResponse<UserResponseDTO>> register(@RequestBody UserBasicDTO userDTO)
+    public ResponseEntity<ApiResponseWrapper<UserResponseDTO>> register(@RequestBody UserBasicDTO userDTO)
             throws RokaMokaContentDuplicated {
         return success(userService.createNormalUser(userDTO));
     }
 
     @PostMapping("/login")
-    public ResponseEntity<ApiResponse<UserResponseDTO>> login(Authentication authentication) {
+    public ResponseEntity<ApiResponseWrapper<UserResponseDTO>> login(Authentication authentication) {
         return success(new UserResponseDTO(authenticationService.authenticate(authentication)));
     }
 }
