+++ src/main/java/br/edu/ufpel/rokamoka/exceptions/GlobalExceptionHandler.java
@@ -1,7 +1,7 @@
 package br.edu.ufpel.rokamoka.exceptions;
 
-import br.edu.ufpel.rokamoka.context.ApiResponse;
-import br.edu.ufpel.rokamoka.wrapper.ApiResponseWrapper;
+import br.edu.ufpel.rokamoka.context.ApiResponseWrapper;
+import br.edu.ufpel.rokamoka.wrapper.RokaMokaController;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -13,10 +13,10 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
  */
 @Slf4j
 @RestControllerAdvice
-public class GlobalExceptionHandler extends ApiResponseWrapper {
+public class GlobalExceptionHandler extends RokaMokaController {
 
     @ExceptionHandler(value = {RokaMokaContentDuplicated.class})
-    public ResponseEntity<ApiResponse<Void>> handleBadRequest(Exception ex) {
+    public ResponseEntity<ApiResponseWrapper<Void>> handleBadRequest(Exception ex) {
         return error(ex, HttpStatus.BAD_REQUEST);
     }
 }
