+++ src/main/java/br/edu/ufpel/rokamoka/wrapper/RokaMokaController.java
@@ -1,51 +1,53 @@
 package br.edu.ufpel.rokamoka.wrapper;
 
-import br.edu.ufpel.rokamoka.context.ApiResponse;
+import br.edu.ufpel.rokamoka.context.ApiResponseWrapper;
 import br.edu.ufpel.rokamoka.context.ServiceContext;
 import br.edu.ufpel.rokamoka.utils.DateUtils;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.RestController;
 
 /**
  * @author mauriciomucci
  */
 @Slf4j
-public abstract class ApiResponseWrapper {
+@RestController
+public abstract class RokaMokaController {
 
     private static final String EXECUTION_UUID = "Execution-UUID";
     private static final String EXECUTION_START_TIME = "Execution-Start-Time";
     private static final String EXECUTION_END_TIME = "Execution-End-Time";
     private static final String EXECUTION_ELAPSED_TIME = "Execution-Elapsed-Time";
 
-    protected ResponseEntity<ApiResponse<Void>> success() {
+    protected ResponseEntity<ApiResponseWrapper<Void>> success() {
         log.info("Sucesso ao executar APIResponse<Void>");
 
         ServiceContext ctx = ServiceContext.getContext();
         ctx.setEndTime(System.currentTimeMillis());
 
         HttpHeaders headers = buildResponseHeaders(ctx);
-        ApiResponse<Void> body = new ApiResponse<>(ctx);
+        ApiResponseWrapper<Void> body = new ApiResponseWrapper<>(ctx);
 
         ServiceContext.clearContext();
         return ResponseEntity.ok().headers(headers).body(body);
     }
 
-    protected <T> ResponseEntity<ApiResponse<T>> success(T response) {
+    protected <T> ResponseEntity<ApiResponseWrapper<T>> success(T response) {
         log.info("Sucesso ao executar APIResponse<{}>", response.getClass().getSimpleName());
 
         ServiceContext ctx = ServiceContext.getContext();
         ctx.setEndTime(System.currentTimeMillis());
 
         HttpHeaders headers = buildResponseHeaders(ctx);
-        ApiResponse<T> body = new ApiResponse<>(ctx, response);
+        ApiResponseWrapper<T> body = new ApiResponseWrapper<>(ctx, response);
 
         ServiceContext.clearContext();
         return ResponseEntity.ok().headers(headers).body(body);
     }
 
-    protected ResponseEntity<ApiResponse<Void>> error(Exception e, HttpStatus status) {
+    protected ResponseEntity<ApiResponseWrapper<Void>> error(Exception e, HttpStatus status) {
         log.error("", e);
 
         ServiceContext ctx = ServiceContext.getContext();
@@ -54,7 +56,7 @@ public abstract class ApiResponseWrapper {
         ctx.setEndTime(System.currentTimeMillis());
 
         HttpHeaders headers = buildResponseHeaders(ctx);
-        ApiResponse<Void> body = new ApiResponse<>(ctx);
+        ApiResponseWrapper<Void> body = new ApiResponseWrapper<>(ctx);
 
         ServiceContext.clearContext();
         return ResponseEntity.status(status).headers(headers).body(body);
